#32 Refactored poll spec into multiple files
