+++ spec/support/nuntium_helper.rb
@@ -1,0 +1,7 @@
+module NuntiumHelper
+
+  def messages
+    @nuntium_ao_messages
+  end
+
+end
